resolution of uss-904 deep linking on deleted (non existing topics) causes an exception http://teamopenuss.uni-muenster.de/jira/browse/USS-904
bugfixed whole deep linking of discussion package.
-> redirects are not executed immediately, but after it was tried to render page.

further exceptions in render phase have to be prevented -> introduced boolean rendered attribute in basebean, which is true, if a redirect was triggered.

there are also checks added, if a topic belongs to a course and if a post belongs to a topic to prevent exploits.

git-svn-id: https://svn.code.sf.net/p/openuss/code/trunk/openuss-plexus@3468 60046477-0c10-0410-bf4b-e552a4ff1768